refactor: avoid duplicated code
